split out FilesItem class into own file
